Check for hosts consistency on list
The hosts file is checked when rendering the nodes list. This allows for
a warning to be issued in the case of discrepancies

The `gethostip` calls need to be done individually. This is because a
single missing node will cause the entire system call to error. Instead
all the system calls are done in parallel