Move the uniq mac check into Models::Node
This means the edit command also preforms the same check. The mac is
stored separately from the params to preform this check, however it is
then merged into to treat it like a "param"

The hunt command needed to be updated to remove the explicit check. It
is no longer possible to reassign nodes across clusters