Handle completely broken symlink in status check
File.exists? returns false a symlink points to a non existent file.
However as the link file still exists, it tricks the status check into
thinking it's in a pending state.

This causes the build to error, as it can not replace the symlink.
Instead, the status check should check for the existance of broken links
and respond accordingly.

PS: also replaced _rendered_path with _path