Use symlink to link groups and nodes together
Finding the groups for a single node is an easy task, as it can directly
check the config.

However reverse looking up the nodes in a group requires `nodeattr`.
This isn't ideal as it makes updating the groups tricky. Instead,
nodeattr has been removed.

In its place, the nodes create symlinks the groups can follow. The link
needs to verified however as they are not automatically deleted. They
are only deleted if they are loaded and found to be invalid