Turn the symlink into a Index file
This symlink isn't required as the target can be determined from the
file path itself. This makes validating the index a bit easier

This commit removes the symlink related code and replaces it with an
Index class. The Index class is similar to a model, but used for this
purpose