Break all the template path methods
These path methods are overly convoluted and tied to the cascade
render. Instead a simplified single render will be used. The source
template will still be defined in a hierarchy, but will be rendered
once against the node to get the final file

A bunch of the commands will be broken because of this. These will be
fixed latter